requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
requests>=2.28.0
python-dotenv>=1.0.0
# vnstock from GitHub (latest) - https://github.com/thinh-vu/vnstock
vnstock @ git+https://github.com/thinh-vu/vnstock.git
yfinance>=0.2.0
websockets>=12.0
